+++ fern/definition/database/enumerate.yml
@@ -8,15 +8,11 @@ types:
   DatabaseEnumerateDetails:
     properties:
       Target: string
+      DatabaseType: DatabaseType
       Banner: optional<string>
       Connected: optional<boolean>
       Fingerprinted: optional<boolean>
       PasswordRequired: optional<boolean>
       ExposedData: optional<list<string>>
       Error: optional<string>
-  DatabaseEnumerateReport:
-    properties:
-      targets: list<string>
-      databaseType: DatabaseType
-      databaseDetails: optional<list<DatabaseEnumerateDetails>>
-      errors: optional<list<string>>
+
