+++ utils/networkapplicationenumerate.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	utilsFern "github.com/Method-Security/networkscan/generated/go/utils"
+	mysql "github.com/Method-Security/networkscan/internal/database/mysql/enumerate"
 	ftp "github.com/Method-Security/networkscan/internal/ftp/enumerate"
 	smtp "github.com/Method-Security/networkscan/internal/smtp/enumerate"
 	ssh "github.com/Method-Security/networkscan/internal/ssh/enumerate"
@@ -113,6 +114,8 @@ func getEngine(application utilsFern.NetworkApplication) (NetworkApplicationEngi
 		return NetworkApplicationEngine{Library: &ftp.LibraryEnumerateFTP{}}, nil
 	case utilsFern.NetworkApplicationSmtp:
 		return NetworkApplicationEngine{Library: &smtp.LibraryEnumerateSMTP{}}, nil
+	case utilsFern.NetworkApplicationMysql:
+		return NetworkApplicationEngine{Library: &mysql.LibraryEnumerate{}}, nil
 	default:
 		return NetworkApplicationEngine{}, fmt.Errorf("unsupported network application: %v", application)
 	}
